Find the resource in the catalog
In order to test that the resource is in the catalog, as we expect it to
be, we perform a search by name. If present, we follow that link and
check for a link to the PURL we have in the configuration for that
resource and return the text of the PURL link.